requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiohttp-retry==2.9.1
aiosignal==1.4.0
amqp==5.3.1
asgiref==3.10.0
attrs==25.4.0
billiard==4.2.2
celery==5.5.3
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
Django==5.2.8
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
frozenlist==1.8.0
gunicorn==23.0.0
idna==3.11
kombu==5.5.4
multidict==6.7.0
packaging==25.0
prompt_toolkit==3.0.52
propcache==0.4.1
psycopg2-binary==2.9.11
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
redis==5.2.1
requests==2.32.5
six==1.17.0
sqlparse==0.5.3
twilio==9.8.6
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
yarl==1.22.0
razorpay==1.3.0
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0
django-redis==5.2.0
dj-database-url==1.0.0
Pillow==10.0.0
Shapely>=1.8.0
python-decouple==3.8
django-cors-headers==4.3.1
drf-spectacular==0.27.1